Added all available entities. FINALLY fixed issue with entities not being highlighted correctly. Now able to move on to other stuff!

+++ ggj14/ggj14/entities/Entity.cs
@@ -32,7 +32,7 @@ namespace ggj14.entities
             return classType;
         }
 
-        public double getFrameCount()
+        public int getFrameCount()
         {
             return numOfFrames;
         }
